refactored test module
